% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_plotly_options.R
\name{clean_plotly_options}
\alias{clean_plotly_options}
\title{Remove Plotly Options on Plots}
\usage{
clean_plotly_options(p)
}
\arguments{
\item{p}{a plotly object}
}
\value{
a plotly object with all options besides 'download as png' on plotly plot
}
\description{
Remove Plotly Options on Plots
}
\examples{
# not run:
# p \%>\%
#   clean_plotly_options()
}
