% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_up.R
\name{round_up}
\alias{round_up}
\title{Round a Whole Number Up}
\usage{
round_up(x, as_dollar = FALSE)
}
\arguments{
\item{x}{a number to round}

\item{as_dollar}{boolean, TRUE to make value with a leading $ and commas where appropiate in the number}
}
\value{
a rounded numeric or a rounded character if as_dollar == TRUE
}
\description{
This was made for reactive UI range to display clean numbers
}
\examples{
round_up(123456789)

round_up(5125, as_dollar = TRUE)
}
